refactor: prune unused code, remove deprecated observability components, and consolidate test modules
