+++ package.yaml
@@ -30,7 +30,6 @@ dependencies:
   - directory >= 1.3.6 && < 1.4
   - resourcet >= 1.2 && < 1.4
   - filepath < 2
-  - ginger >= 0.10.6 && < 0.11
   - http-client
   - http-client-tls
   - http-conduit >= 2 && < 3
@@ -51,7 +50,6 @@ dependencies:
   - servant-conduit >= 0.16 && < 0.17
   - servant-event-stream >= 0.4 && < 0.5
   - sqlite-simple >= 0.4.18 && < 0.5
-  - stache >= 2.3 && < 2.5
   - stm >= 2.5 && < 2.6
   - tagsoup < 0.15
   - text >= 1.2 && < 3
