+++ src/Langchain/Prelude.hs
@@ -200,12 +200,6 @@ module Langchain.Prelude
   , dispatchEventAsync
   , newLoggingCallbackHandler
 
-    -- * Config Validation
-  , ConfigIssue (..)
-  , ValidationResult (..)
-  , validateLangchainConfig
-  , assertValidConfig
-
     -- * Resilience
   , CircuitState (..)
   , CircuitBreakerConfig (..)
@@ -248,7 +242,7 @@ module Langchain.Prelude
     -- * Embeddings
   , Embeddings (..)
 
-    -- * Document Loaders & Transformers
+    -- * Document Loaders
   , Document (..)
   , BaseLoader (..)
   , FileLoader (..)
@@ -264,11 +258,6 @@ module Langchain.Prelude
   , defaultHtmlLoader
   , WebPageLoader (..)
   , defaultWebPageLoader
-  , DocumentTransformer (..)
-  , enrichDocumentMetadata
-  , enrichDocuments
-  , MetadataEnricher (..)
-  , newMetadataEnricher
 
     -- * Prompt Templates
   , PromptTemplate (..)
@@ -324,7 +313,7 @@ module Langchain.Prelude
   , newMapReduceChain
   , runMapReduceChain
 
-    -- * Structured Output, Routers, and Advanced Parsers
+    -- * Structured Output & Parsers
   , OutputParser (..)
   , CommaSeparatedList (..)
   , JSONOutputStructure (..)
@@ -338,16 +327,6 @@ module Langchain.Prelude
   , withJsonFormat
   , withSchemaFormat
   , withStructuredOutput
-  , SemanticRouter (..)
-  , newSemanticRouter
-  , Route (..)
-  , routeQuery
-  , XmlOutputParser (..)
-  , newXmlOutputParser
-  , parseXmlOutput
-  , EnumParser (..)
-  , newEnumParser
-  , parseEnum
 
     -- * Agents & Execution
   , ReActAgent (ReActAgent)
@@ -378,33 +357,6 @@ module Langchain.Prelude
   , LLMReranker (..)
   , newLLMReranker
 
-    -- * Vector Store Metadata Filtering & Header Injection
-  , FilterPredicate (..)
-  , evalFilter
-  , filterDocuments
-  , eqFilter
-  , inFilter
-  , andFilter
-  , orFilter
-  , HeaderInjector (..)
-  , newHeaderInjector
-  , injectChunkHeader
-  , injectChunkHeaders
-
-    -- * Dynamic Flows & Event Streaming Protocol
-  , FlowNode (FlowNode)
-  , FlowEdge (..)
-  , DynamicFlow (..)
-  , FlowExecutionResult (..)
-  , NodeExecutor
-  , ComponentRegistry
-  , topologicalSortFlow
-  , executeDynamicFlow
-  , newDynamicFlow
-  , AgentStreamEvent (..)
-  , formatSSE
-  , formatNdJson
-
     -- * Providers
   , Ollama (..)
   , OllamaClientConfig (..)
@@ -433,7 +385,6 @@ import Langchain.Cache.Core
 import Langchain.Callback.Manager
 import Langchain.Chain.MapReduce
 import Langchain.Chain.RetrievalQA
-import Langchain.Config.Validation
 import Langchain.Core.Error
 import Langchain.Core.Model
 import Langchain.Core.Monad hiding (defaultConfig)
@@ -449,30 +400,23 @@ import Langchain.DocumentLoader.Html
 import Langchain.DocumentLoader.Json
 import Langchain.DocumentLoader.PdfLoader
 import Langchain.DocumentLoader.WebPage
-import Langchain.DocumentTransformer.HeaderInjector
-import Langchain.DocumentTransformer.MetadataEnricher
 import Langchain.Embeddings.Core
 import Langchain.Graph.Blackboard
 import Langchain.Graph.Checkpointer
 import Langchain.Graph.Debate
-import Langchain.Graph.DynamicFlow
 import Langchain.Graph.HITL
 import Langchain.Graph.MultiAgent
 import Langchain.Graph.Parallel
 import Langchain.Graph.StateGraph
 import Langchain.Graph.Voting
 import Langchain.Guardrail.Core
-import Langchain.Logging.Structured
 import Langchain.MCP.Client
 import Langchain.Memory.Core
 import Langchain.Memory.Entity
 import Langchain.Memory.Summary
-import Langchain.Observability.OpenTelemetry
-import Langchain.Observability.StreamProtocol
+import Langchain.Observability
 import Langchain.OutputParser.Core
-import Langchain.OutputParser.Enum
 import Langchain.OutputParser.Structured
-import Langchain.OutputParser.Xml
 import Langchain.Pipeline.DSL
 import Langchain.PromptTemplate.FewShot
 import Langchain.PromptTemplate.Prompt
@@ -504,7 +448,6 @@ import Langchain.Retriever.Hybrid
 import Langchain.Retriever.MultiQueryRetriever
 import Langchain.Retriever.ParentDocument
 import Langchain.Retriever.Reranker
-import Langchain.Router.Semantic
 import Langchain.TextSplitter.Character
 import Langchain.TextSplitter.Code
 import Langchain.TextSplitter.Markdown
@@ -513,7 +456,6 @@ import Langchain.TextSplitter.Token
 import Langchain.Tool.Async
 import Langchain.Tool.GenericSchema
 import Langchain.VectorStore.Core
-import Langchain.VectorStore.Filter
 import Langchain.VectorStore.InMemory
 import Langchain.VectorStore.PgVector
 import Langchain.VectorStore.SqliteVec
