+++ test/Test/Langchain/OutputParser/AdvancedParsersSpec.hs
@@ -13,20 +13,13 @@ import Data.Aeson (FromJSON, ToJSON, Value (..))
 import qualified Data.Aeson.KeyMap as KM
 import Data.Proxy (Proxy (..))
 import Data.Text (Text)
-import qualified Data.Text as T
-import qualified Data.Text.Lazy as TL
 import qualified Data.Vector as V
 import GHC.Generics (Generic)
 import Test.Tasty
 import Test.Tasty.HUnit
 
 import Langchain.Core.Model (userMessage)
-import Langchain.DocumentLoader.Core (Document (..))
-import Langchain.Embeddings.Core (Embeddings (..))
-import Langchain.OutputParser.Enum
 import Langchain.OutputParser.Structured
-import Langchain.OutputParser.Xml
-import Langchain.Router.Semantic
 import Test.Langchain.Provider.Mock (newMockModel)
 
 data TestPerson = TestPerson
@@ -50,21 +43,6 @@ data TestCompany = TestCompany
   , companyEmployees :: [TestPerson]
   }
   deriving (Show, Eq, Generic, ToJSON, FromJSON, StructuredOutput)
-
-data Sentiment = Positive | Negative | Neutral
-  deriving (Show, Eq, Enum, Bounded)
-
-data RouterMockEmbeddings = RouterMockEmbeddings
-
-instance Embeddings RouterMockEmbeddings where
-  embedDocuments _ docs = pure $ map (embedOne . TL.toStrict . pageContent) docs
-  embedQuery _ q = pure $ embedOne q
-
-embedOne :: Text -> [Float]
-embedOne t
-  | "billing" `T.isInfixOf` T.toLower t || "invoice" `T.isInfixOf` T.toLower t = [1.0, 0.0, 0.0]
-  | "tech" `T.isInfixOf` T.toLower t || "bug" `T.isInfixOf` T.toLower t = [0.0, 1.0, 0.0]
-  | otherwise = [0.0, 0.0, 1.0]
 
 tests :: TestTree
 tests =
@@ -107,30 +85,4 @@ tests =
           Just ollamaS -> do
             let rt = fromOllamaSchema ollamaS
             toOllamaSchema rt @?= Just ollamaS
-    , testCase "SemanticRouter routes query to best semantic route" $ do
-        let emb = RouterMockEmbeddings
-            routes =
-              [ Route "billing" "Billing and invoices" ["Where is my invoice?", "Billing question"]
-              , Route "tech_support" "Technical support and bugs" ["I found a bug in the code", "Tech issue"]
-              ]
-            router = newSemanticRouter emb routes 0.8
-        resBilling <- runExceptT $ routeQuery router "I have a question about my invoice"
-        resBilling @?= Right (Just "billing")
-        resTech <- runExceptT $ routeQuery router "There is a bug in the system"
-        resTech @?= Right (Just "tech_support")
-    , testCase "XmlOutputParser parses XML tags into typed structure" $ do
-        let xmlText = "<response><title>Haskell Guide</title><category>FP</category></response>"
-            parser = newXmlOutputParser ["title", "category"] $ \elems ->
-              case (lookupTag "title" elems, lookupTag "category" elems) of
-                (Just t, Just c) -> Right (t, c)
-                _ -> Left "Missing title or category"
-            lookupTag tag es = case [xmlContent e | e <- es, xmlTag e == tag] of
-              (x : _) -> Just x
-              [] -> Nothing
-        parseXmlOutput parser xmlText @?= Right ("Haskell Guide", "FP")
-    , testCase "EnumParser parses fuzzy case-insensitive enum options" $ do
-        (parseEnum "positive" :: Either Text Sentiment) @?= Right Positive
-        (parseEnum "NEGATIVE" :: Either Text Sentiment) @?= Right Negative
-        (parseEnum "  Neutral  " :: Either Text Sentiment) @?= Right Neutral
-        (parseEnum "The sentiment is positive." :: Either Text Sentiment) @?= Right Positive
     ]
