+++ test/Test/Langchain/PromptTemplate/PromptSpec.hs
@@ -47,35 +47,6 @@ tests =
           renderPrompt
             (fromTemplate "Hello, {name:{width}}!")
             (Map.fromList [("name", "Alice"), ("width", "10")])
-    , testCase "renders mustache sections" $
-        let mustacheTemplate = fromTemplateWithFormat "{{#name}}Hello, {{name}}!{{/name}}" Mustache Map.empty
-            mustacheVars = Map.singleton "name" "Alice"
-         in renderPrompt mustacheTemplate mustacheVars @?= Right "Hello, Alice!"
-    , testCase "returns an error for missing mustache variables" $
-        let mustacheTemplate = fromTemplateWithFormat "Hello, {{name}}!" Mustache Map.empty
-         in case renderPrompt mustacheTemplate Map.empty of
-              Left err -> "name" `T.isInfixOf` T.pack (show err) @? "Expected error to contain 'name'"
-              Right _ -> assertFailure "Expected an error for missing mustache variable"
-    , testCase "renders jinja2 conditionals" $
-        let jinjaTemplate =
-              fromTemplateWithFormat
-                "{% if enabled %}Hello, {{ name }}!{% else %}Disabled{% endif %}"
-                Jinja2
-                Map.empty
-            jinjaVars = Map.fromList [("enabled", "true"), ("name", "Alice")]
-         in renderPrompt jinjaTemplate jinjaVars @?= Right "Hello, Alice!"
-    , testCase "renders jinja2 filters" $
-        let jinjaTemplate = fromTemplateWithFormat "Hello, {{ name | upper }}!" Jinja2 Map.empty
-            jinjaVars = Map.singleton "name" "Alice"
-         in renderPrompt jinjaTemplate jinjaVars @?= Right "Hello, ALICE!"
-    , testCase "infers jinja2 variables from conditionals" $
-        let jinjaTemplate = fromTemplateWithFormat "{% if enabled %}Hello, {{ name }}!{% endif %}" Jinja2 Map.empty
-         in inputVariables jinjaTemplate @?= ["enabled", "name"]
-    , testCase "returns an error for missing jinja2 variables" $
-        let jinjaTemplate = fromTemplateWithFormat "Hello, {{ name }}!" Jinja2 Map.empty
-         in case renderPrompt jinjaTemplate Map.empty of
-              Left err -> "name" `T.isInfixOf` T.pack (show err) @? "Expected error to contain 'name'"
-              Right _ -> assertFailure "Expected an error for missing jinja2 variable"
     ]
   where
     greetingTemplate = fromTemplate "Hello, {name}! Welcome to {place}."
