--- test/data/prompt_file.txt
@@ -1,2 +1,0 @@
-Question: {question}
-Answer:
