--- test/fixtures/gemini_embed_response.json
@@ -1,5 +1,0 @@
-{
-  "embedding": {
-    "values": [0.0123, -0.0456, 0.0789]
-  }
-}
